refactor: Untemplate Generic::LayerBuilder
