+++ Examples/Detectors/GenericDetector/CMakeLists.txt
@@ -4,6 +4,7 @@ add_library(
     src/BuildGenericDetector.cpp
     src/GenericDetector.cpp
     src/GenericDetectorElement.cpp
+    src/LayerBuilder.cpp
 )
 
 acts_compile_headers(ActsExamplesDetectorGeneric GLOB "include/**/*.hpp")
